+++ src/providers/myNotesProvider.ts
@@ -397,7 +397,7 @@ export class MyNotesProvider implements vscode.TreeDataProvider<TreeNode> {
       item.command = {
         command: 'hackmd.ui.edit',
         title: 'Open Note',
-        arguments: [{ type: 'note', note, preserveFocus: true }],
+        arguments: [note, { preserveFocus: true }],
       };
     } else {
       item.command = undefined;