+++ src/providers/teamNotesProvider.ts
@@ -512,7 +512,7 @@ export class TeamNotesProvider implements vscode.TreeDataProvider<TreeNode> {
       item.command = {
         command: 'hackmd.ui.edit',
         title: 'Open Note',
-        arguments: [{ type: 'note', note, preserveFocus: true }],
+        arguments: [note, { preserveFocus: true }],
       };
     } else {
       item.command = undefined;